refactor: dao patterns and interfaces for service beans

+++ src/main/java/com/atcollabo/hackathon/classdojo/service/impl/TeacherServiceImpl.java
@@ -34,7 +34,7 @@ public class TeacherServiceImpl implements TeacherService {
 
     @Override
     public int getTotalClassSessions(Long classId) {
-        return teacherDao.getTotalClassSessions(classId);
+        return (int)teacherDao.getTotalClassSessions(classId);
     }
 
     @Override
